first-commit

+++ service/index.js
@@ -4,24 +4,28 @@ const createContact = (body) => {
   return Contact.create(body);
 };
 
-const getContacts = () => {
-  return Contact.find();
+const getContacts = (id) => {
+  return Contact.find({ owner: id });
 };
 
-const getContactById = (id) => {
-  return Contact.findOne({ _id: id });
+const getContactById = (id, owner) => {
+  return Contact.findOne({ _id: id, owner });
 };
 
-const updateContact = (id, body) => {
-  return Contact.findByIdAndUpdate({ _id: id }, body, { new: true });
+const updateContact = (id, owner, body) => {
+  return Contact.findOneAndUpdate({ _id: id, owner }, body, { new: true });
 };
 
-const removeContact = (id) => {
-  return Contact.findByIdAndRemove({ _id: id });
+const removeContact = (id, owner) => {
+  return Contact.findByIdAndRemove({ _id: id, owner });
 };
 
-const updateFavoriteContact = (id, favorite) => {
-  return Contact.findByIdAndUpdate(id, { favorite }, { new: true });
+const updateFavoriteContact = (id, owner, favorite) => {
+  return Contact.findOneAndUpdate(
+    { _id: id, owner },
+    { favorite },
+    { new: true }
+  );
 };
 
 module.exports = {
